requirements.txt
discord.py
groq
requests
python-dotenv